Separate core library in different files
