+++ .gitignore
@@ -1,3 +1,4 @@
 build
 dist
 *.egg-info
+**/*.pyc
