+++ pytorch_grad_cam/guided_backprop.py
@@ -1,0 +1,70 @@
+import numpy as np
+import torch
+from torch.autograd import Function
+
+
+class GuidedBackpropReLU(Function):
+    @staticmethod
+    def forward(self, input_img):
+        positive_mask = (input_img > 0).type_as(input_img)
+        output = torch.addcmul(torch.zeros(input_img.size()).type_as(input_img), input_img, positive_mask)
+        self.save_for_backward(input_img, output)
+        return output
+
+    @staticmethod
+    def backward(self, grad_output):
+        input_img, output = self.saved_tensors
+        grad_input = None
+
+        positive_mask_1 = (input_img > 0).type_as(grad_output)
+        positive_mask_2 = (grad_output > 0).type_as(grad_output)
+        grad_input = torch.addcmul(torch.zeros(input_img.size()).type_as(input_img),
+                                   torch.addcmul(torch.zeros(input_img.size()).type_as(input_img), grad_output,
+                                                 positive_mask_1), positive_mask_2)
+        return grad_input
+
+
+class GuidedBackpropReLUModel:
+    def __init__(self, model, use_cuda):
+        self.model = model
+        self.model.eval()
+        self.cuda = use_cuda
+        if self.cuda:
+            self.model = model.cuda()
+
+        def recursive_relu_apply(module_top):
+            for idx, module in module_top._modules.items():
+                recursive_relu_apply(module)
+                if module.__class__.__name__ == 'ReLU':
+                    module_top._modules[idx] = GuidedBackpropReLU.apply
+
+        # replace ReLU with GuidedBackpropReLU
+        recursive_relu_apply(self.model)
+
+    def forward(self, input_img):
+        return self.model(input_img)
+
+    def __call__(self, input_img, target_category=None):
+        if self.cuda:
+            input_img = input_img.cuda()
+
+        input_img = input_img.requires_grad_(True)
+
+        output = self.forward(input_img)
+
+        if target_category is None:
+            target_category = np.argmax(output.cpu().data.numpy())
+
+        one_hot = np.zeros((1, output.size()[-1]), dtype=np.float32)
+        one_hot[0][target_category] = 1
+        one_hot = torch.from_numpy(one_hot).requires_grad_(True)
+        if self.cuda:
+            one_hot = one_hot.cuda()
+
+        one_hot = torch.sum(one_hot * output)
+        one_hot.backward(retain_graph=True)
+
+        output = input_img.grad.cpu().data.numpy()
+        output = output[0, :, :, :]
+
+        return output
